premgr

+++ src/arduino/motor-controller/MotorWrapper.h
@@ -14,22 +14,20 @@ public:
   void setMaxSpeed(float value) override { stepper.setMaxSpeed(value); }
   void setAcceleration(float value) override { stepper.setAcceleration(value); }
   void setSettingsFlags(uint8_t value) override { _settingsFlags = value; }
-  // void setScriptActive(bool value) override { _scriptActive = value; }
-  void setScriptActive(bool value) override {  }
-  void setDelayActive(bool value) override {  }
+  void setScriptActive(bool value) override { }
+  void setDelayActive(bool value) override { }
   void setSleepTimeoutMs(uint16_t value) override { _sleepTimeoutMs = value; }
   void setEnableDelayMs(uint8_t value) override { _enableDelayMs = value; }
   uint8_t stateFlags() override { return _stateFlags; }
   uint8_t settingsFlags() override { return _settingsFlags; }
-  int32_t currentPosition() override { return (int32_t) stepper.currentPosition(); }
-  int32_t targetPosition() override { return (int32_t) stepper.targetPosition(); }
+  int32_t currentPosition() override { return stepper.currentPosition(); }
+  int32_t targetPosition() override { return stepper.targetPosition(); }
   float speed() override { return (float) abs(stepper.speed()); }
   float maxSpeed() override { return (float) stepper.maxSpeed(); }
   float acceleration() override { return (float) stepper.acceleration(); }
   uint16_t sleepTimeoutMs() override { return _sleepTimeoutMs; }
   uint8_t enableDelayMs() override { return _enableDelayMs; }
   bool busy() override { return stepper.isRunning(); }
-  // bool scriptActive() override { return _scriptActive; }
 
 protected:
   AccelStepper& stepper;
@@ -38,8 +36,6 @@ protected:
   uint16_t _sleepTimeoutMs = 2000;
   uint8_t _enableDelayMs = 2;
 
-// private:
-//   bool _scriptActive = false;
 };
 
 #endif
